requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.9.0
aiosqlite==0.20.0
openai==1.50.0
numpy==2.1.0
scipy==1.14.0
